# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управління транспортними засобами (автомобілі, бренди, категорії тощо).",
        "title": "GoVehicleApi",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/vehicle/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle"
                ],
                "summary": "Список транспортних засобів",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер сторінки",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Кількість на сторінці",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vehicle.GetAllResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle"
                ],
                "summary": "Створити транспортний засіб",
                "parameters": [
                    {
                        "description": "Дані транспортного засобу",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vehicle.VehiclePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vehicle.SaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/vehicle/expanded/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicle"
                ],
                "summary": "Отримати розгорнутий транспортний засіб за ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID транспортного засобу",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vehicle.GetExpandedSwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "body_style": {
                    "type": "integer"
                },
                "brand": {
                    "type": "integer"
                },
                "category": {
                    "type": "integer"
                },
                "driver_type": {
                    "type": "integer"
                },
                "gearbox": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "vehicle.ExpandedVehicleSwagger": {
            "type": "object",
            "properties": {
                "body_style": {
                    "$ref": "#/definitions/vehicle.swaggerBodyStyle"
                },
                "brand": {
                    "$ref": "#/definitions/vehicle.swaggerCompressedBrand"
                },
                "category": {
                    "$ref": "#/definitions/vehicle.swaggerCategory"
                },
                "driver_type": {
                    "$ref": "#/definitions/vehicle.swaggerDriverType"
                },
                "gearbox": {
                    "$ref": "#/definitions/vehicle.swaggerGearbox"
                },
                "id": {
                    "type": "integer"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "vehicle.GetAllResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "$ref": "#/definitions/response.Response"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vehicle"
                    }
                }
            }
        },
        "vehicle.GetExpandedSwaggerResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "$ref": "#/definitions/response.Response"
                },
                "vehicle": {
                    "$ref": "#/definitions/vehicle.ExpandedVehicleSwagger"
                }
            }
        },
...
        "vehicle.SaveResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "$ref": "#/definitions/response.Response"
                },
                "vehicle": {
                    "$ref": "#/definitions/models.Vehicle"
                }
            }
        },
...
        "vehicle.VehiclePayload": {
            "type": "object",
            "properties": {
                "Vehicle": {
                    "type": "object",
                    "properties": {
                        "body_style": {
                            "type": "integer"
                        },
                        "brand": {
                            "type": "integer"
                        },
                        "category": {
                            "type": "integer"
                        },
                        "driver_type": {
                            "type": "integer"
                        },
                        "gearbox": {
                            "type": "integer"
                        },
                        "mileage": {
                            "type": "integer"
                        },
                        "model": {
                            "type": "string"
                        },
                        "price": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "vehicle.swaggerBodyStyle": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "vehicle.swaggerCategory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "vehicle.swaggerCompressedBrand": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "vehicle.swaggerDriverType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "vehicle.swaggerGearbox": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введіть токен у форматі: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}